Reorganized Guard Statuses
