+++ src/fighter/common/common_status.rs
@@ -4,7 +4,10 @@ mod run_brake;
 mod jump_squat;
 mod jump;
 mod tread_jump;
-mod shield;
+mod guard_on;
+mod guard;
+mod guard_off;
+mod guard_damage;
 mod escape;
 mod escape_air;
 mod passive;
@@ -37,7 +40,10 @@ pub fn install() {
     jump_squat::install();
     jump::install();
     tread_jump::install();
-    shield::install();
+    guard_on::install();
+    guard::install();
+    guard_off::install();
+    guard_damage::install();
     escape::install();
     escape_air::install();
     passive::install();
