+++ src/fighter/common/common_status/damage.rs
@@ -2,7 +2,6 @@ use {
     smash::{
         lua2cpp::{L2CFighterCommon, *},
         hash40,
-        phx::*,
         app::{lua_bind::*, *},
         lib::{lua_const::*, L2CValue}
     },
